:fsn_inventory: Bug/Exploit fix
Before fix: If you click drop on an item, exit out, then go back in and drop the same or any other item with any quantity it will drop both the original item as well as the new item. In Apartments you could drop and store items.

This sometimes glitched out and caused the items to be duped in quantity. "Vesomplay" came to me last night regarding this issue, letting me know how it was done/recreated so it could be fixed.

Added "UpdateOnscreenKeyboard() == 2" 2 is User has canceled editing.